requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
arrow==0.15.7
astroid==2.4.2
asyncio==3.4.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bidict==0.22.1
bs4==0.0.1
cachelib==0.9.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.0.4
cheroot==9.0.0
click==8.1.3
colorama==0.4.4
configparser==5.0.2
cors==1.0.1
coverage==6.4.1
crayons==0.4.0
cryptowatch-sdk==0.1.0
cssselect==1.1.0
dateparser==1.1.1
dnspython==2.2.1
docker-pycreds==0.4.0
fake-useragent==0.1.11
feedparser==6.0.10
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SocketIO==5.3.2
future==0.18.2
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.30
gpio==0.3.0
greenlet==2.0.1
idna==2.10
importlib-metadata==4.12.0
isort==4.3.21
itsdangerous==2.1.2
jaraco.functools==3.5.2
Jinja2==3.1.2
lazy-object-proxy==1.4.3
lxml==4.9.1
MarkupSafe==2.1.1
marshmallow==3.13.0
mccabe==0.6.1
more-itertools==9.0.0
Naked==0.1.32
numpy==1.23.3
pandas==1.5.0
pandoc-latex-fontsize==1.1.3
panflute==2.2.3
parse==1.19.0
pathtools==0.1.2
plumbum==1.7.2
protobuf==3.20.1
psutil==5.9.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycryptodome==3.16.0
pyee==8.2.2
pylint==2.5.3
pymongo==4.3.3
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.1
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.3.2
requests==2.25.1
requests-file==1.5.1
requests-html==0.10.0
rpyc==5.1.0
selenium==3.141.0
sentry-sdk==1.14.0
setproctitle==1.3.2
sgmllib3k==1.0.0
shellescape==3.8.1
six==1.15.0
smmap==5.0.0
soupsieve==2.3.2.post1
tldextract==3.4.0
toml==0.10.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.2
tzlocal==4.2
ujson==5.7.0
unicorn-binance-websocket-api==1.41.0
unicorn-fy==0.12.2
urllib3==1.26.14
w3lib==2.0.1
wandb==0.13.9
webdriver-manager==3.4.2
websocket-client==1.4.2
websockets==10.3
Werkzeug==2.2.2
wrapt==1.12.1
yliveticker==0.3.0
zipp==3.8.1
zope.event==4.6
zope.interface==5.5.2
